+++ classes/suggested-tasks/class-local-tasks-manager.php
@@ -9,12 +9,15 @@ namespace Progress_Planner\Suggested_Tasks;
 
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Local_Task_Factory;
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Content_Create;
-use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Content_Update;
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Content_Review;
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Core_Update;
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Core_Blogdescription;
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Settings_Saved;
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Yoast_Organization_Logo;
-
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Debug_Display;
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Sample_Page;
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Hello_World;
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Core_Siteicon;
 
 /**
  * Local_Tasks_Manager class.
@@ -46,10 +49,14 @@ class Local_Tasks_Manager {
 
 		$this->task_providers = [
 			new Content_Create(),
-			new Content_Update(),
+			new Content_Review(),
 			new Core_Update(),
 			new Core_Blogdescription(),
 			new Settings_Saved(),
+			new Debug_Display(),
+			new Sample_Page(),
+			new Hello_World(),
+			new Core_Siteicon(),
 		];
 
 		\add_filter( 'progress_planner_suggested_tasks_items', [ $this, 'inject_tasks' ] );
@@ -92,13 +99,13 @@ class Local_Tasks_Manager {
 	/**
 	 * Get a task provider by its type.
 	 *
-	 * @param string $provider_type The provider type.
+	 * @param string $provider_id The provider ID.
 	 *
 	 * @return \Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Local_Tasks_Interface|null
 	 */
-	public function get_task_provider( $provider_type ) {
+	public function get_task_provider( $provider_id ) {
 		foreach ( $this->task_providers as $provider_instance ) {
-			if ( $provider_instance->get_provider_type() === $provider_type ) {
+			if ( $provider_instance->get_provider_id() === $provider_id ) {
 				return $provider_instance;
 			}
 		}
@@ -160,7 +167,7 @@ class Local_Tasks_Manager {
 	 */
 	public function evaluate_task( $task_id ) {
 		$task_object   = ( new Local_Task_Factory( $task_id ) )->get_task();
-		$task_provider = $this->get_task_provider( $task_object->get_provider_type() );
+		$task_provider = $this->get_task_provider( $task_object->get_provider_id() );
 
 		if ( ! $task_provider ) {
 			return false;
@@ -179,7 +186,7 @@ class Local_Tasks_Manager {
 
 		foreach ( $tasks as $task ) {
 			$task_object = ( new Local_Task_Factory( $task ) )->get_task();
-			$provider    = $this->get_task_provider( $task_object->get_provider_type() );
+			$provider    = $this->get_task_provider( $task_object->get_provider_id() );
 			if ( ! $provider ) {
 				$this->remove_pending_task( $task );
 			}
@@ -195,7 +202,7 @@ class Local_Tasks_Manager {
 	 */
 	public function get_task_details( $task_id ) {
 		$task_object   = ( new Local_Task_Factory( $task_id ) )->get_task();
-		$task_provider = $this->get_task_provider( $task_object->get_provider_type() );
+		$task_provider = $this->get_task_provider( $task_object->get_provider_id() );
 
 		if ( ! $task_provider ) {
 			return false;
@@ -283,8 +290,18 @@ class Local_Tasks_Manager {
 				$task_object = ( new Local_Task_Factory( $task ) )->get_task();
 				$task_data   = $task_object->get_data();
 
+				// If the task was already completed, remove it.
+				if ( true === \progress_planner()->get_suggested_tasks()->was_task_completed( $task_data['task_id'] ) ) {
+					return false;
+				}
+
 				if ( isset( $task_data['year_week'] ) ) {
 					return \gmdate( 'YW' ) === $task_data['year_week'];
+				}
+
+				// We have changed type name, so we need to remove all tasks of the old type.
+				if ( isset( $task_data['type'] ) && 'update-post' === $task_data['type'] ) {
+					return false;
 				}
 
 				return true;